+++ plugins/imap/go.mod
@@ -1,0 +1,16 @@
+module imap
+
+go 1.25.1
+
+require (
+	github.com/emersion/go-imap/v2 v2.0.0-beta.6
+	github.com/emersion/go-message v0.18.2
+	github.com/wombat v0.0.0-00010101000000-000000000000
+)
+
+require (
+	github.com/emersion/go-sasl v0.0.0-20241020182733-b788ff22d5a6 // indirect
+	golang.org/x/text v0.29.0 // indirect
+)
+
+replace github.com/wombat => ../../
